Split out DRM and ioctl
