+++ ioctl_iface.c
@@ -1,0 +1,18 @@
+#include <linux/types.h>
+
+#include "ioctl_iface.h"
+
+int ioctl_probe(void)
+{
+	return 0;
+}
+
+void ioctl_remove(void)
+{
+	return;
+}
+
+void ioctl_shutdown(void)
+{
+	return;
+}
