+++ ioctl_iface.h
@@ -1,0 +1,8 @@
+#ifndef IOCTL_IFACE_H_
+#define IOCTL_IFACE_H_
+
+int ioctl_probe(void);
+void ioctl_remove(void);
+void ioctl_shutdown(void);
+
+#endif
