+++ params_iface.h
@@ -4,4 +4,8 @@
 extern int g_param_mono_cutoff;
 extern int g_param_mono_invert;
 
+int params_probe(void);
+void params_remove(void);
+void params_shutdown(void);
+
 #endif
